Don't retry sockets with timeout set to 0
For non-blocking sockets with the timeout set to 0, the select call in
_recv and _send will return immediately without the socket being ready.
This causes _recv to return None, which makes recv close the connection.
Since a select with a timeout of 0 right after recv/send is pointless,
just skip _recv and _send when timeout is 0.

As for the handling of SSLWantReadError, SSLWantWriteError, EAGAIN and
EWOULDBLOCK in _recv/_send, websocket-client can't handle these without
blocking. Since applications which has timeout set to 0 don't want
websocket-client to block them, they have to handle those errors
themselves.

Fixes #535