wsdump: Fix --headers option
Since https://github.com/websocket-client/websocket-client/pull/506
changes the way that options['header'] is used, it breaks --headers
option of wsdump command because wsdump command uses a map object
instead of a list here, which can only be used once. Therefore, after
the first use with 'not in' operator, it becomes empty and no headers
specified by the user are sent.

To fix it, simply convert the map object into a list, so it can be use
multiple times.