Revert #417 and reimplement proxy support with python-socks
Due to the ongoing license change process in #697, PR #417 from @jhtitor
needed to be removed. PR #417 added SOCKS proxy support using
PySocks, which is the most popular Python SOCKS proxy and is
used by libraries such as 'requests' and 'urllib3'. However, as of
mid-2021, PySocks has not seen any activity since Sep 2019 and has
several open issues. Because PySocks does not appear to be maintained
and has no forks that are actively trying to replace the old
repository, python-socks may be a better choice going forward. Although
python-socks is a newer project, it has good code coverage and recent commits from earlier this year (2021).