BREAKING: Add url parameter to set Origin header
Fixes #787. Most interactions with a HTTPS website use a wss://
WebSocket, and the initial WebSocket request usually uses a https:// URL
in the Origin header. Previously, a http:// Origin header was the
default value unless a custom origin was specified. Now the Origin header will dynamically set a https:// URL
in the Origin header when a wss:// WebSocket URL is passed instead of
defaulting to http://. This can be
overridden by setting a custom "origin" parameter.